+++ DependencyInjection/Configuration.php
@@ -45,6 +45,7 @@ class Configuration implements ConfigurationInterface
             'stack_exchange',
             'vkontakte',
             'windows_live',
+            'wordpress',
             'yandex',
             '37signals',