+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 * Added Box resource owner
 * Added Dailymotion resource owner
 * Added DeviantArt resource owner
+* Added Wordpress resource owner
 
 ## 0.3.0-alpha2 (2013-07-29)
 * [BC break] Added `ResourceOwnerInterface::isCsrfTokenValid()` method