+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-## 0.2.7 (2013-xx-xx)
+## 0.2.7 (2013-08-03)
 * Fix: Polish oauth error detection to cover cases from i.e. Facebook resource owner
 * Fix: Changed authorization url for Vkontakte resource owner